cmd: add custom testnet flags (#2715)
Adds custom testnet flags to `charon create cluster` and `charon run` commands.

category: feature
ticket: #2690

+++ cluster/eip712sigs.go
@@ -80,7 +80,7 @@ func digestEIP712(typ eip712Type, def Definition, operator Operator) ([]byte, er
 		Domain: eip712.Domain{
 			Name:    "Obol",
 			Version: "1",
-			ChainID: uint64(chainID),
+			ChainID: chainID,
 		},
 		Type: eip712.Type{
 			Name: typ.PrimaryType,
